Renamed MemoryUtils to PcapUtils.
